tested email verification
