--- index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Speak Hello World</title>
-</head>
-<body>
-    <button onclick="speak()">Click to Speak</button>
-
-    <script>
-        function speak() {
-            const utterance = new SpeechSynthesisUtterance("Hello World");
-            const synth = window.speechSynthesis;
-            const voices = synth.getVoices();
-            
-            // Find the "Microsoft Jenny Online (Natural) - English (United States)" voice
-            const voice = voices.find(voice => voice.voiceURI === "Microsoft Jenny Online (Natural) - English (United States)");
-            
-            if (voice) {
-                utterance.voice = voice;
-            } else {
-                alert("Voice not found. Using the default voice instead.");
-            }
-
-            window.speechSynthesis.speak(utterance);
-        }
-
-        // Ensure voices are loaded
-        window.speechSynthesis.onvoiceschanged = function() {
-            window.speechSynthesis.getVoices();
-        };
-    </script>
-</body>
-</html>
