--- templates/emailVerified.html
@@ -1,40 +1,0 @@
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles3.css') }}">
-  <link rel="icon" type="image/x-icon" href="{{ url_for('static', filename='images/logo.png') }}">
-    <title>Email Verified</title>
-    <style>
-        h1 {
-            color: black;
-            font-family: Arial, sans-serif;
-            font-size: 1.5em;
-            text-align: center;
-        }
-        a {
-            color: blue;
-            text-decoration: none;
-        }
-        a:hover {
-            color: darkblue;
-            text-decoration: underline;
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-      <div class="content-box">
-
-
-            <h1>Email has been verified!</h1>
-            <h1>Check your inbox! Email should appear soon.</h1>
-            <br>
-            <h1>Thanks for using CivicConnect. Your support means a lot!</h1>
-            <h1>Use again? Click <a href="/">here</a></h1>
-      </div>
-      </div>
-
-  
- 
-  </html>
