+++ templates/end.html
@@ -55,7 +55,7 @@
     <div class="form-container">
 
       <h1 class="header" id="civicConnect">Thank you!</h1>
-      <p>Thank you so much for using CivicConnect! The email should appear in your inbox soon and you have been CC'd on the email. It may take some time for the email to appear, we are still starting off!</p>
+      <p>Thank you so much for using CivicConnect! The email should appear in your inbox soon and you have been CC'd on the email. It may take some time for the email to appear, as we are still starting off!</p>
 
       <p>Want to use the program again? Then click here: <a href="/exit">HOME</a></p>
